relocate StreamScope to utils / quarks.streamscope
