+++ build.xml
@@ -112,6 +112,7 @@
         <ant dir="runtime/jsoncontrol" target="@{target}" useNativeBasedir="true"/>
         <ant dir="runtime/jobregistry" target="@{target}" useNativeBasedir="true"/>
         <ant dir="utils/metrics" target="@{target}" useNativeBasedir="true"/>
+        <ant dir="utils/streamscope" target="@{target}" useNativeBasedir="true"/>
 
         <ant dir="providers/direct" target="@{target}" useNativeBasedir="true"/>
         <ant dir="connectors/common" target="@{target}" useNativeBasedir="true"/>
@@ -215,6 +216,7 @@
       </classfiles>
     <sourcefiles encoding="UTF-8">
         <fileset dir="${basedir}/utils/metrics/src/main/java" includes="**/*.java"/>
+        <fileset dir="${basedir}/utils/streamscope/src/main/java" includes="**/*.java"/>
     </sourcefiles>
     </group>
     <group name="Quarks Connectors">
@@ -321,7 +323,7 @@
        <group title="Quarks Connectors" packages="quarks.connectors.*"/>
        <group title="Quarks Samples" packages="quarks.samples.*"/>
        <group title="Quarks Analytics" packages="quarks.analytics.*"/>
-       <group title="Quarks Utilities" packages="quarks.metrics,quarks.metrics.*"/>
+       <group title="Quarks Utilities" packages="quarks.metrics,quarks.metrics.*,quarks.streamscope,quarks.streamscope.*"/>
        <group title="Quarks Low-Level API" packages="quarks.graph,quarks.graph.*,quarks.oplet,quarks.oplet.*,quarks.window"/>
        <group title="Quarks SPI" packages="quarks.*.spi.*"/>
      </javadoc>
