+++ utils/.classpath
@@ -2,6 +2,8 @@
 <classpath>
 	<classpathentry kind="src" path="metrics/src/main/java"/>
 	<classpathentry kind="src" path="metrics/src/test/java"/>
+	<classpathentry kind="src" path="streamscope/src/main/java"/>
+	<classpathentry kind="src" path="streamscope/src/test/java"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/api"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/ext"/>
