+++ utils/streamscope/build.xml
@@ -14,27 +14,24 @@
  specific language governing permissions and limitations
  under the License.
  -->
-<project name="quarks.providers.development" default="all"
-    xmlns:jacoco="antlib:org.jacoco.ant"
-    >
+<project name="quarks.utils.streamscope" default="all" xmlns:jacoco="antlib:org.jacoco.ant">
     <description>
-        Build the development topology provider.
+        Build the StreamScope add-on.
     </description>
 
+  <property name="component.path" value="utils/streamscope"/>
   <import file="../../common-build.xml"/>
 
   <path id="compile.classpath">
-    <pathelement location="${quarks.lib}/quarks.providers.direct.jar" />
-    <pathelement location="${quarks.console}/server/lib/quarks.console.server.jar" />
-    <pathelement location="${quarks.utils}/metrics/lib/quarks.utils.metrics.jar" />
-    <pathelement location="${quarks.utils}/streamscope/lib/quarks.utils.streamscope.jar" />
-    <pathelement location="${lib}/quarks.runtime.jmxcontrol.jar" />
+    <pathelement location="${quarks.lib}/quarks.api.topology.jar"/>
+    <path refid="quarks.ext.classpath"/>
   </path>
 
   <path id="test.compile.classpath">
     <pathelement location="${jar}" />
-    <pathelement location="../../api/topology/test.classes" />
-    <pathelement location="../../utils/streamscope/test.classes" />
+    <pathelement location="${quarks.lib}/quarks.providers.direct.jar"/>
+    <pathelement location="../../api/topology/test.classes"/>
+    <pathelement location="../../providers/direct/test.classes"/>
     <path refid="compile.classpath"/>
   </path>
