+++ utils/streamscope/src/main/java/quarks/streamscope/StreamScopeRegistry.java
@@ -16,7 +16,7 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-package quarks.topology.plumbing;
+package quarks.streamscope;
 
 import java.util.ArrayList;
 import java.util.Collections;
